chore: add pyproject.toml and prepare for publish to pypi
