+++ README.md
@@ -26,8 +26,40 @@ This project is created to support Cambodian culture and make Khmer calendar log
 
 ## 🚀 Usage
 
-### 1. Clone the Repository
+```shell
+# from pypi
+pip install khmerdate
 
-```bash
-git clone https://github.com/khmer-lunar-date/khmer_lunar_date.git
-cd khmer_lunar_date
+# or directly from source
+pip install git+https://github.com/danhhong/khmer_lunar_date.git
+```
+
+---
+
+```python
+import datetime
+from khmerdate import gregorian_to_khmer_lunar, khmer_day_of_week
+
+day, month, year = 1, 7, 2025
+
+result = gregorian_to_khmer_lunar(day, month, year)
+print(result)
+
+# =>
+{
+    "lunar_day": "៦កើត",
+    "lunar_month": "អាសាឍ",
+    "lunar_year": "២៥៦៩",
+    "zodiac_year": "ម្សាញ់",
+    "stem": "សប្តស័ក",
+}
+
+# formatting as string
+day_name = datetime.datetime(year, month, day).strftime("%A")
+day_name = khmer_day_of_week[day_name]
+
+result_fmt = f"ថ្ងៃ{day_name} {result['lunar_day']} ខែ{result['lunar_month']} ឆ្នាំ{result['zodiac_year']} {result['stem']} ព.ស. {result['lunar_year']}"
+
+print(result_fmt)
+# => ថ្ងៃអង្គារ ៦កើត ខែអាសាឍ ឆ្នាំម្សាញ់ សប្តស័ក ព.ស. ២៥៦៩
+```
