+++ pyproject.toml
@@ -1,0 +1,21 @@
+
+[project]
+name = "khmerdate"
+version = "0.1.0"
+authors = [{ name = "Danh Hong", email = "danhhong@gmail.com" }]
+description = "Khmer Lunar Date Converter"
+readme = "README.md"
+
+classifiers = [
+    "Development Status :: 5 - Production/Stable",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3",
+    "Topic :: Scientific/Engineering :: Artificial Intelligence",
+]
+
+[project.urls]
+repository = "https://github.com/danhhong/khmer_lunar_date"
+
+[tool.setuptools.packages.find]
+where = ["khmerdate"]
